fix(kernel,ata,virtio-blk): make the barrier and the shutdown survive real media
Adds the test that proves a WFS volume survives a boot, and fixes the six bugs
writing it exposed. Every one of them was invisible to the existing suites by
construction: the WFS drive runs with snapshot=on, so guest writes land in an
overlay that is discarded, and a cache flush against an overlay returns
instantly. The suite could not observe durability, which is the one property the
journal barriers exist to provide.

QemuConfig.wfs_snapshot turns the overlay off for a test that owns a scratch copy
of the image. tests/test_wfs_clean_unmount.py is the only such test.

THE HALTING PROCESS STARVED THE PROCESS MANAGER. kernel_system_shutdown armed
the request and then YIELDED in a loop waiting for the PM to step the sequence.
On a single-CPU guest that keeps the caller runnable, so the scheduler has no
reason to schedule the PM at all: the sequence never started, no participant was
notified, and the machine powered off through the stall fallback as though the
PM had wedged. Measured 0 of 4 boots reaching the sequence; parking on an
endpoint nothing sends to (ipc_endpoint_wait_for) makes it 4 of 4. run-qemu-test
could not see this -- it halts once per run and passed 8 of 8 with the markers
present, because a single boot happens to schedule the PM before the fallback
expires.

ATA CACHE FLUSH WAITED ON THE BUDGET SIZED FOR A SECTOR TRANSFER. Every other
command this driver issues moves at most a few sectors; a flush commits the
drive's entire write cache and is bounded by how much dirty data it holds. The
driver gave up while the drive was still BSY and issued its next command into a
desynchronised channel, hanging every request behind it -- 302 seconds measured
against a 13 second baseline. With its own budget it is 13.4 seconds.

Two further defects in that same budget, both caught in review:

- The standalone BLOCK_IPC_FLUSH_REQ path -- the journal barrier itself, and the
  reason any of this exists -- did not get the new budget. Only the flush ending
  a write did. The measurement improved anyway, because a write path's own
  flushes dominate, which is how a half-fix hides.
- The budget was keyed on whether the interrupt is live, but ata_wait_step only
  SLEEPS on the primary channel, that being the only line routed. The WFS volume
  is unit 2, secondary, so it spun through the sleep-sized budget and the large
  poll budget was unreachable. The improvement was 200 -> 3000 iterations, not
  the 50000x the constant implied, and the next slower disk would regress. It is
  now keyed on whether a wait step sleeps.

ata_flush_unit also gains the 400 ns settle after the drive select, and an ERR
check after the flush completes. That check exists here and nowhere else in the
driver because a flush is the one command whose failure a caller cannot detect
any other way: a barrier reporting success over a failed flush lets the writer
put its COMMIT block on media believing the images it names are already there.

THE CLEAN MARK WAS NOT BARRIERED. wfs_sync_task issued no flush, which ATA
masked by ending every write with one; on virtio-blk the WFS_STATE_CLEAN
superblock could sit in a volatile write cache at exactly the moment power is
cut. It is now barriered like a journal step -- the one write whose entire
purpose is to be read by the next mount.

THE PARTICIPANT DEADLINE WAS SHORTER THAN THE WORK IT WAITS ON. At 1 s, a
fs-wfs sync -- superblock plus backups, each write ending in a flush -- could
overrun, and the next thing the sequence does when the last participant is passed
over is cut the power. A deadline shorter than the write it interrupts turns a
clean unmount into a torn superblock, which is worse than never having run. Now
8 s, which costs nothing on the normal path where the participant answers in
milliseconds.

A FLUSH AGED AGAINST VIRTIO-BLK'S TRANSFER WATCHDOG RESET THE DEVICE. Expiry
there is not a failed request: quiesce() resets the device and errors everything
outstanding, so a flush that was merely slow cost the filesystem its disk for the
rest of the boot. A flush now has its own deadline, for the same reason the ATA
side does.

Smaller, all from the same review: pm_service_set replaced a service's flags on
re-registration, so a service that declared WANTS_SHUTDOWN through the descriptor
path silently dropped out of the sequence by re-registering its name (they
accumulate now); a park that failed to park would have degenerated into the spin
the change exists to remove, with no diagnostic; and the halting CPU's reads of
g_pm.shutdown.active/.index now use the pm_atomic_* helpers, with the header's
invariant corrected to say those two are read cross-CPU.

The test asserts across boots in two classes and four boots. A clean halt must
leave the volume CLEAN with its bytes on the media; a KILLED machine must not
leave it looking cleanly unmounted. The control is not optional: without it the
clean assertion passes against a driver that prints that line unconditionally, or
one that never marks the volume dirty. It asserts "not clean" rather than
"replays" -- a kill after the writer finished leaves the volume dirty with an
EMPTY log, every transaction having been checkpointed and retired, so demanding a
replay would demand the kill land inside a transaction, which a test cannot time.

fs_wfs logs which of the three mount paths a volume took, because the state byte
never leaves the driver and "mounted" alone cannot tell a clean unmount from a
replay that happened to find nothing.

What the test covers of the flush bug is the flush ending a WRITE, which is what
a guest write path issues; the standalone barrier shares the budget but is not
separately provoked, since nothing lets a test make one flush slow on demand.
Four boots is four times the exposure to the boot wedge in docs/TASKS.md, and
that is the dominant reason this file will fail when it fails -- one wedge in
eight probe boots. The test itself passed 4 of 4 runs.

Verified: run-kernel-unit-tests, run-qemu-test (all three shutdown markers),
fmt-check, lint, --verify-batteries, tests.test_wfs_mount_read +
tests.test_wfs_virtio_blk (18), and tests.test_wfs_clean_unmount.

+++ src/kernel/include/process_manager_internal.h
@@ -157,14 +157,18 @@ typedef struct {
 
 /* The orderly shutdown sequence's state (process_manager_shutdown.c).
  *
- * `requested` is the only field written from another CPU -- the halting
+ * `requested` is the only field WRITTEN from another CPU -- the halting
  * process's, through kernel_system_shutdown_arm -- and is accessed with the
  * pm_atomic_* helpers for that reason. Everything else is written by
- * pm_shutdown_step on the PM's own CPU, after it has observed that store. */
+ * pm_shutdown_step on the PM's own CPU, after it has observed that store.
+ *
+ * `active` and `index` are READ from the halting process's CPU as well, by the
+ * wait loop in kernel_system_shutdown, so they are accessed with the helpers
+ * too. The pair is read as one forward-only quantity and never written there. */
 typedef struct {
     uint32_t requested; /* armed by the halt/reboot host call */
     uint32_t reason;    /* WASMOS_SHUTDOWN_REASON_* */
-    uint8_t active;     /* participants have been collected */
+    uint32_t active;    /* participants have been collected; read cross-CPU */
     uint8_t pending;    /* the participant at `index` has been notified */
     uint32_t count;
     uint32_t index; /* the participant being waited on */
